Update native UltraNativeModuleTraceEvent with uuid
